% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mult.R
\name{mult}
\alias{mult}
\title{Multiply together two numbers}
\usage{
mult(4,5)
}
\arguments{
\item{x}{A number}

\item{y}{Another number}
}
\value{
product The product of multiplying x and y
}
\description{
Multiply together two numbers
}
